Change classes Extendable and Controller, adapting files
Exposed actions now has to end with the word 'Action',
for example: indexAction, loginAction, etc... Because of
this new structure, the list of exposed actions is no longer
needed as so the isExposedMethod. The routes is also adapted
they only has to check if the requested action is a function
and fallback to the default action.